+++ core/templates/collection/create.ts
@@ -28,7 +28,7 @@ export default async function template(
   return `
 ${
     breadcrumb(options, version, [
-      collection.name,
+      collection.label,
       getPath(basePath, "collection", collection.name),
     ], "New file")
   }